In the previous implementation, the frame sent to the system contained the elapsed (#8)
time for a given _tick_. However, only systems running at the same frame rate as the
game loop have an elapsed time identical to the tick time. If we run physics at 1FPS
and the elapsed time between game loop ticks is 100ms, then the elapsed time on the frame
sent to physics should be 1000ms (time between physics frames, not time between game
loop ticks).

So the fix for this is to multiply the elapsed time from the game loop tick by the system's
modulus (number of ticks per frame). So physics with a framerate of 1fps has a modulus
of 10 (10 gameloop ticks per frame) when the game loop is running at 10fps, this
makes the elapsed time sent to physics 100ms * 10 = 1000ms = 1s, which is the
"real elapsed time" between physics frames.